Initial version
This is the initial version of the nanomsg distributor, created for the Lumith ECM system. It is unchanged from it's late 2017 creation time with the exception that the paths have been tweaked to make it sensible for distribution and the Makefile changed to build statically.

Copyright of this commit remains David Helkowski.