Reworked distributor into base for orchestration system
Reworked the 'distributor' ( nanodist.c which is now norch.c ) into the basic beginning of an agent that takes commands from a director and responds to it with results.

The agent also listens to broadcast communication from the director.

The director receives results from agents, it can send out broadcast communication to all agents, and it can respond to direct requests.

The client is the CLI for controlling the director. It makes requests of the director via request/reply.

In this state the groundwork is laid and it is close to being ready to use structured communication via XJR between the parts and make the agent actually start doing useful things.

Note that currently there is no encryption or signing between any of the components. This is obviously not acceptable for production use, but is left this way for the time being as this whole setup is largely a proof of concept.

The options being considered for later:
* Use NNG instead of nanomsg, and the TLS layer it provides ( seems heavyweight to me )
* Add in an ECC symmetric encrpyption library and use pre-shared key between all nodes ( this is lightweight and fast; will likely be the solution )
* Add in some sort of public key encryption ( once again, heavyweight, but people like this stuff )
* Add in some TLS library ( this is a bad option; might as well move to NNG )