Modularize director
Created different modules / parts for the director so that each threaded portion is in its own file for more clarity.

The datastore is for storing the source of truth for registered nodes / config etc.

The results part is for storing the results of items ( what actual things to be done by an agent/node are being called )

The scheduler collects new items, and sends them out to agents/nodes to be done.

Reqrep part handles requests incoming to the director, mainly from clients.

This commit is in a broken state in the middle of still moving things around and implementing the needed logic to actually make a new item from a client, pass to the director, get it scheduled, send out to a agent/node, run, and results send back to the director.