Implement request/item scheduling
Requests/Items should now be able to be added, scheduled, and sent to agents. A result req type is stubbed out in the results part, and there is a stub of a 'cmd' item type within the agent.

The code as a whole is still not tested in its current state, so it likely still retains some minor bugs.