Simplify code and item notation. Implement item_cmd for agent
Moved to using node names for the type of items, rather than them being 'req' nodes with a type attribute. This simplifies the code a bunch. Changed 'req' to 'item' in a bunch of places to use a single way to refer to lineitems in a series of commands being done.

Removed un-needed hash wrapper on returns from execution of various commands.

Made it possible for threads to process multiple items per block of data sent in through nanomsg, with some exceptions ( an id is not currently assigned to each item as it should be; TODO )

Implemented execution and capture of stdout, stderr, and errorlevel of arbitrary commands run through an agent. Signals are accepted and dealt with but not properly captured yet. The results of the command is formed as a message but not yet send back to the results part of the director. TODO next.