Modularized agent. Fixed director bugs.
Within item_cmd, check to see if cmd file requested to be executed even exists and is marked executable and fail early if not.

Change item_* function style so that it returns the result directly, so that each item_* does not need to handle sending results back.

Put error decoding helper into a misc file so it can be shared by portions of the agent.

Split agent into multiple threads so that sending results, running items, and accepting incoming items are all handled asynchronously and won't block each other. ( not yet complete; but nearly there )

Fix minor bugs spread through the director making it non-functional. Also changed the director to use TCP sockets between forks... cannot use inproc between forks. Should change to using perl threads, and see if it works. It may not though, and definitely wouldn't work in non-threaded perl. Leaving it alone for now as the whole thing is intended to be rewritten in C; once C it can properly be done with threads.

Adding various debugging print statements through the director. Will need to drop them or put them behind a debugging flag later. TODO

Added gitignore for test_cmd so that it doesn't show up in git status.