Make results returning work, fix bugs, simplify socket creation
Implemented the portions of the agent returner so that results can in theory be returned to the director. It looks though currently that the results aren't actually being returned and somehow the item itself is being sent back oddly instead of the results.

Added various debugging print statements all over; weirdly they don't seem to always print out.

Simplified socket creation because there was duplicated code for it all over and getting messy.

Changed the inproc nanomsg to tcp sockets for runner and returner of agent, in attempt to see if it reduced random failures I was seeing. It does not seem to make any difference but leaving it that way for the moment.

Am currently seeing timeout when attempting to return results to the director from the agent. Not quite sure why; is a bug. TODO fix this.