Split out config of agent. Fix bugs prevent results returning.
Put all config reading into its own file storing results of reading config into a struct for ease of use and clean handling of config.

Also added a function to print out config for debugging.

Added a debugging output to the misc make_nn_socket function so that it can be easily seen what is binded or connected nanomsg-wise. This revealed the runner and returned were not being bound. Corrected that and returning results and overall operation now works reasonably instead of randomly.

TODO: Add build flags to indicate that functions are not being used as that would have revealed the issue also.