Get item dependencies working
It is now possible to set a single 'dep' on a item, to point at a specifically named item. ( item with 'name' attribute )

The client allows items to be wrapped in a inorder tag to automatically assign dep and name to the series of items within it.

Added a dummy 'output' item that just lets the scheduler output a part of a template directly, in preparation for combining the current code with perl-Template-Nepl to allow variables to be used by subsequents items.