Passing variables between items along with templated items
Added Template::Nepl usage into the system for items to operate based on the results of items before them.

In the process added the ability to process items asynchronously to the datastore. Originally the datastore was always direct request/reply but this became a deadlock issue in the case of template evaluation due to datastore calling scheduler, in turn calling back to datastore.

Began converting debugging output statements to state the part and function they are output from for better ease is seeing what is going on.

Encountered issues with recursion of cdata begin/end tags passing things around. For now am escaping them with ||>, but this will only work one level deep. To resolve correctly it will need to be corrected within Parse::XJR.

Encountered issues with not holding onto the Parse::XJR root of an item. Once the root goes out of scope all nodes queried out of it will be using freed memory. Fixed it temporarily for now by holding onto the root. This also should be resolved in Parse::XJR by way of the addition of a 'root here' function that duplicates a portion of the tree away from the original root, allowing the original root to go free.

Am ocasionally seeing the agent itself crash; have yet invesigated exactly why. The agent needs to be run through valgrind and appropriate memory frees added where needed and/or correct any use after frees. It is likely that use after free is causing the crashes. It is also possible that the crashes only occurred due to garbage cdata begin/end tags being sent into it, so it may not be anything serious.

Am storing results of items being run, and they are accessible now but in a non-clear way. TODO: Make it easier and more clear on how to accessible the results of named items and/or the previous item run in a series of items.

TODO: Add a 'foreach' item type, to loop over all agents of a specific role. TODO: Add a 'doall' item, to run all items under it asynchronously instead of in order.